# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wild Rift Macro Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ModelInfoPricing = {
  prompt: string;
  completion: string;
};

export interface ModelInfo {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  context_length?: number | null;
  pricing?: ModelInfoPricing;
}

export interface GameContext {
  /**
     * Current game time (e.g. "8:42")
     * @nullable
     */
  gameTime?: string | null;
  /**
     * Player location on map (e.g. "Mid lane", "Dragon pit", "Jungle")
     * @nullable
     */
  myLocation?: string | null;
  /**
     * Player role (e.g. "Mid", "ADC", "Support", "Jungle", "Top")
     * @nullable
     */
  myRole?: string | null;
  /**
     * Comma-separated ally champion names
     * @nullable
     */
  allyChampions?: string | null;
  /**
     * Comma-separated enemy champion names
     * @nullable
     */
  enemyChampions?: string | null;
  /**
     * Dragon status (e.g. "Up", "Down 1:30", "Soul point")
     * @nullable
     */
  dragonStatus?: string | null;
  /**
     * Baron status (e.g. "Up", "Down 2:00")
     * @nullable
     */
  baronStatus?: string | null;
  /**
     * Rift Herald status
     * @nullable
     */
  riftHeraldStatus?: string | null;
  /**
     * Gold difference (e.g. "+500", "-1200")
     * @nullable
     */
  goldDiff?: string | null;
  /**
     * Kill score (e.g. "8-5")
     * @nullable
     */
  score?: string | null;
  /**
     * Any additional context
     * @nullable
     */
  additionalNotes?: string | null;
}

export interface AnalyzeInput {
  /**
     * Base64 encoded screenshot image
     * @nullable
     */
  imageBase64?: string | null;
  /**
     * Base64 encoded annotated minimap image showing player positions
     * @nullable
     */
  minimapBase64?: string | null;
  /** OpenRouter model ID to use */
  model: string;
  context?: GameContext;
}

export interface OpenrouterConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenrouterMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenrouterConversationInput {
  title: string;
  model: string;
}

export interface OpenrouterConversationPatch {
  title?: string;
}

export interface OpenrouterMessageInput {
  content: string;
  model: string;
  context?: GameContext;
}

export interface OpenrouterConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenrouterMessage[];
}

export interface OpenrouterError {
  error: string;
}