# lib/api-zod/src/generated/types/openrouterConversationPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wild Rift Macro Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenrouterConversationPatch {
  title?: string;
}